suppress errors and fix if condition

+++ lib/oc_clean_istags.sh
@@ -51,8 +51,8 @@ mapfile -t normalized_branches <<< "$(git branch -r | sed 's/origin\///g' | tr '
 
 for istag in "${istags[@]}"
 do
-    if [ -z "$($git tag --points-at "$istag")" ]; then # never delete imagestream tag if there is a git tag pointing at it
-        if [[ "$($git cat-file -t "$istag" 2>/dev/null)" != "commit" || ! ${normalized_branches[*]} =~ $istag ]]; then
+    if [ -z "$($git tag --points-at "$istag" 2>/dev/null)" ]; then # never delete imagestream tag if there is a git tag pointing at it
+        if [[ "$($git cat-file -t "$istag" 2>/dev/null)" != "commit" && ! ${normalized_branches[*]} =~ $istag ]]; then
             # Delete tags pointing to objects that do not exist anymore
             # both commits and branches return the type "commit"
             $oc -n "$oc_project" delete istag/"$image_stream":"$istag"
